# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Lychee API
  description: |
    Lychee is a universal local LLM runtime with multi-protocol API support.
    It extends Ollama with Anthropic API compatibility, Model Composer, Agent Sandbox,
    and an embedded Web Dashboard.
  version: 0.1.0-alpha
  license:
    name: MIT
    url: https://github.com/MD-Mushfiqur123/lychee/blob/main/LICENSE
  contact:
    name: MD-Mushfiqur123
    url: https://github.com/MD-Mushfiqur123/lychee
servers:
  - url: http://localhost:11434
    description: Local Lychee server (default)
...
paths:
...
  /api/structured:
    post:
      tags: [structured]
      summary: Generate schema-conforming JSON with auto-retry
      operationId: generateStructured
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, prompt]
              properties:
                model:
                  type: string
                prompt:
                  type: string
                schema:
                  type: object
                max_retries:
                  type: integer
                  default: 3
                options:
                  type: object
      responses:
        '200':
          description: Structured generation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
                  valid:
                    type: boolean
                  attempts:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: string
...
  /api/routes/{name}/status:
    get:
      tags: [routes]
      summary: Query the health and status of a virtual model route
      operationId: getRouteStatus
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Route status details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
...
    ModelEndpointStatus:
      type: object
      required: [host, healthy, active_requests]
      properties:
        host:
          type: string
          example: "http://localhost:11434"
        model:
          type: string
          example: "gemma3"
        healthy:
          type: boolean
          example: true
        active_requests:
          type: integer
          example: 0
        last_check:
          type: string
          format: date-time
        last_error:
          type: string
          example: "connection refused"
    RouteStatusResponse:
      type: object
      required: [name, strategy, endpoints]
      properties:
        name:
          type: string
          example: "fast"
        strategy:
          type: string
          example: "round_robin"
        endpoints:
          type: array
          items:
            $ref: '#/components/schemas/ModelEndpointStatus'
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'